% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tree.R
\name{fetch_tree}
\alias{fetch_tree}
\title{Returns the mosaic database}
\usage{
fetch_tree(TreeURL)
}
\arguments{
\item{TreeURL}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
