% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.load.package.R
\name{install.load.package}
\alias{install.load.package}
\title{Returns the mosaic database}
\usage{
install.load.package(x)
}
\arguments{
\item{x}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
